Update test code coverage
