+++ ClickHouse.Client/ADO/ClickHouseCancelableConnectionFactory.cs
@@ -4,7 +4,7 @@ using ClickHouse.Client.ADO.Parameters;
 
 namespace ClickHouse.Client.ADO;
 
-#if NET5_0_OR_GREATER
+#if NET7_0_OR_GREATER
 public class ClickHouseCancelableConnectionFactory : DbProviderFactory
 {
     public static ClickHouseCancelableConnectionFactory Instance => new();
@@ -19,8 +19,6 @@ public class ClickHouseCancelableConnectionFactory : DbProviderFactory
 
     public override DbCommand CreateCommand() => new ClickHouseCancelableCommand();
 
-#if NET7_0_OR_GREATER
     public override DbDataSource CreateDataSource(string connectionString) => new ClickHouseDataSource(connectionString);
-#endif
 }
 #endif
